Rename Bible model to Bible:Translation
